Refine iterator seeking behavior with respect to boundaries
When seeking an iterator, we’re looking for the earliest line or token
that starts at the given position or ends after it. That is, we’ll seek
to an *empty* line or token ending at the position, but otherwise favor
a token starting at the position.

The exception is with the last token on a line or the last line in the
index, which can end at the target position when seeking.

This also drops the return values from TokenIterator’s seek methods,
since they weren’t needed and were complex to support.